refactor: rename google_auth -> drive_auth for consistency
